Refactor again. Use lists instead of hashmaps.

+++ reporting/cit-energy-project-reporting/src/main/java/de/tuberlin/cit/project/energy/reporting/model/PowerHistoryEntry.java
@@ -2,9 +2,9 @@ package de.tuberlin.cit.project.energy.reporting.model;
 
 public class PowerHistoryEntry extends HistoryEntry {
     private final String hostname;
-    private final double usedPower;
+    private final float usedPower;
 
-    public PowerHistoryEntry(long timestamp, String hostname, double usedPower) {
+    public PowerHistoryEntry(long timestamp, String hostname, float usedPower) {
         super(timestamp);
         this.hostname = hostname;
         this.usedPower = usedPower;
@@ -14,7 +14,7 @@ public class PowerHistoryEntry extends HistoryEntry {
         return hostname;
     }
     
-    public double getUsedPower() {
+    public float getUsedPower() {
         return usedPower;
     }
 }
